SNAP-709: ERRORCODE=-4470, SQLSTATE=08003
