+++ src/main/resources/jobs-cms-hibernate.cfg.xml
@@ -61,6 +61,7 @@
 		<property name="hibernate.c3p0.debugUnreturnedConnectionStackTraces">true</property>
 
 		<property name="hibernate.c3p0.testConnectionOnCheckout">true</property>
+		<property name="hibernate.c3p0.connectionTesterClassName">gov.ca.cwds.neutron.quark.NeutronNoOpConnectionTester</property>
 		<property name="hibernate.c3p0.preferredTestQuery">SELECT 1 FROM SYSIBM.SYSDUMMY1 FOR READ ONLY WITH UR</property>
 
 	</session-factory>
